can write data and read data

+++ C++/sfs/client.h
@@ -33,7 +33,7 @@ class Client
 
         int64_t get_length(uint64_t file_id);
 
-        int32_t read(uint64_t file_id, char* buffer, int64_t &length);
+        int64_t read(uint64_t file_id, char* buffer, int64_t &length);
         int32_t save_metadata();
         
         // for test
